+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to
 
 ### Fixed
 
+- Replace the "not working?" prompt by "All good, but if your credential stops
+  working, you may need to re-authorize here.".
+  [#2102](https://github.com/OpenFn/lightning/issues/1872)
+
 ## [v2.5.0-pre2] - 2024-05-17
 
 ### Changed